refactor(review): isolate closed epoch accounting
